Remove the DOF containers from the thermal-mechanics functional since they are on longer used
The physics sub-modules (solid mechanics and thermal conduction) hold the DOFs.
Add getters in the thermal-mechanics functional that just forwards the DOFS from the
sub-modules.

Paired with Sam Mish.

+++ src/serac/physics/solid_functional.hpp
@@ -178,11 +178,11 @@ public:
     mesh_.NewNodes(*mesh_nodes, true);
   }
 
-  void setParameter(const FiniteElementState & parameter_state, int i)
+  void setParameter(const FiniteElementState& parameter_state, size_t i)
   {
     functional_call_args_[i + 1] = parameter_state.trueVec();
   }
- 
+
   /**
    * @brief Set essential displacement boundary conditions (strongly enforced)
    *
